Add lval_pop and lval_take
lval_pop is a function that takes a single element from an S-Expression at index  and shifts the list backwards so lval* is no longer present. It returns the extracted value. lval_take deletes the list it extracted the element from. Both  make it possible to evaluate the S-expression and children.